// src/lib/types.ts
export interface Skill {
  name: string;
  description: string;
  path: string;
  baseDir: string;
  scope: 'global' | 'workspace';
}

export interface CloudSkill {
  rank: number;
  slug: string;
  name: string;
  description: string;
  version: string;
  homepage: string;
  downloads: number;
  stars: number;
  score: number;
  categories: string[];
}

export interface Workflow {
  name: string;
  description: string;
  path: string;
  workspace?: string;
  content?: string;
  scope?: 'global' | 'workspace';
  baseDir?: string;
}

export interface Rule {
  name: string;
  description: string;
  path: string;
  content?: string;
  scope?: 'global' | 'workspace';
  baseDir?: string;
}